refactor: split QUANTUMSTRAND into first-class layout/tags modules
Remove the floss/qs product package and qs CLI/binary packaging. Move
layout analysis into floss/layout, string tagging into floss/tags (with
DBs still under floss/qs/db/data), and extract FLOSS argparse into
floss/cli. Promote former [qs] dependencies to core install deps.
