--- .github/workflows/build-qs.yml
@@ -1,118 +1,0 @@
-name: build quantumstrand
-
-on:
-  pull_request:
-    branches: [ quantumstrand ]
-  push:
-    branches: [ quantumstrand ]
-  release:
-    types: [edited, published]
-    branches: [ quantumstrand ]
-
-jobs:
-  build:
-    name: PyInstaller for ${{ matrix.os }}
-    runs-on: ${{ matrix.os }}
-    strategy:
-      matrix:
-        include:
-          # use old linux so that the shared library versioning is more portable
-          - os: ubuntu-22.04
-            artifact_name: quantumstrand
-            asset_name: linux
-          - os: windows-2022
-            artifact_name: quantumstrand.exe
-            asset_name: windows
-          - os: macos-14
-            artifact_name: quantumstrand
-            asset_name: macos
-    # Pin action version by commit hash to maximize trust, ref: https://securitylab.github.com/research/github-actions-building-blocks/
-    steps:
-      - name: Checkout floss
-        uses: actions/checkout@11bd71901bbe5b1630ceea73d27597364c9af683 # v4.2.2
-        with:
-          submodules: false
-          lfs: true
-      - name: Set up Python 3.11
-        uses: actions/setup-python@a26af69be951a213d495a4c3e4e4022e16d87065 # v5.6.0
-        with:
-          python-version: '3.11'
-      - name: Install floss[build]
-        run: pip install -e .[build]
-      - name: Install floss[qs]
-        run: pip install -e .[qs]
-      - name: Build standalone executable
-        run: pyinstaller .github/pyinstaller/qs.spec
-      - uses: actions/upload-artifact@ea165f8d65b6e75b540449e92b4886f43607fa02 # v4.6.2
-        with:
-          name: ${{ matrix.asset_name }}
-          path: dist/${{ matrix.artifact_name }}
-
-  test_run:
-    name: Test run on ${{ matrix.os }}
-    runs-on: ${{ matrix.os }}
-    needs: [build]
-    strategy:
-      matrix:
-        include:
-          - os: ubuntu-22.04
-            artifact_name: quantumstrand
-            asset_name: linux
-          - os: windows-2022
-            artifact_name: quantumstrand.exe
-            asset_name: windows
-          - os: macos-14
-            artifact_name: quantumstrand
-            asset_name: macos
-    steps:
-      - name: Download ${{ matrix.asset_name }}
-        uses: actions/download-artifact@d3f86a106a0bac45b974a628896c90dbdf5c8093 # v4.3.0
-        with:
-          name: ${{ matrix.asset_name }}
-      - name: Checkout testfiles
-        uses: actions/checkout@11bd71901bbe5b1630ceea73d27597364c9af683 # v4.2.2
-        with:
-          repository: mandiant/flare-floss-testfiles
-          path: tests/data
-      - name: Set executable flag
-        if: startsWith(matrix.os, 'windows') != true
-        run: chmod +x ${{ matrix.artifact_name }}
-      - name: Run quantumstrand -h
-        run: ./${{ matrix.artifact_name }} -h
-      - name: Run quantumstrand
-        run: ./${{ matrix.artifact_name }} tests/data/test-decode-to-stack.exe
-
-  zip_and_upload:
-    # upload zipped binaries to Release page
-    if: github.event_name == 'release'
-    name: zip and upload ${{ matrix.asset_name }}
-    runs-on: ubuntu-22.04
-    needs: [build]
-    permissions:
-      contents: write
-    strategy:
-      matrix:
-        include:
-          - asset_name: linux
-            artifact_name: quantumstrand
-          - asset_name: windows
-            artifact_name: quantumstrand.exe
-          - asset_name: macos
-            artifact_name: quantumstrand
-    steps:
-      - name: Download ${{ matrix.asset_name }}
-        uses: actions/download-artifact@d3f86a106a0bac45b974a628896c90dbdf5c8093 # v4.3.0
-        with:
-          name: ${{ matrix.asset_name }}
-      - name: Set executable flag
-        run: chmod +x ${{ matrix.artifact_name }}
-      - name: Set zip name
-        run: echo "zip_name=quantumstrand-${GITHUB_REF#refs/tags/}-${{ matrix.asset_name }}.zip" >> $GITHUB_ENV
-      - name: Zip ${{ matrix.artifact_name }} into ${{ env.zip_name }}
-        run: zip ${{ env.zip_name }} ${{ matrix.artifact_name }}
-      - name: Upload ${{ env.zip_name }} to GH Release
-        uses: svenstaro/upload-release-action@81c65b7cd4de9b2570615ce3aad67a41de5b1a13 # 2.11.2
-        with:
-          repo_token: ${{ secrets.GITHUB_TOKEN }}
-          file: ${{ env.zip_name }}
-          tag: ${{ github.ref }}
