+++ floss/tags/gp.py
@@ -8,7 +8,12 @@ from dataclasses import dataclass
 
 import msgspec
 
-import floss.qs.db
+
+def _data_root() -> pathlib.Path:
+    """Legacy DB location: floss/qs/db/data."""
+    return pathlib.Path(__file__).resolve().parents[1] / "qs" / "db" / "data"
+
+
 
 Encoding = Literal["ascii"] | Literal["utf-16le"] | Literal["unknown"]
 # header | gap | overlay
@@ -126,11 +131,11 @@ class StringHashDatabase:
 
 
 DEFAULT_PATHS = (
-    pathlib.Path(floss.qs.db.__file__).parent / "data" / "gp" / "gp.jsonl.gz",
-    pathlib.Path(floss.qs.db.__file__).parent / "data" / "gp" / "cwindb-native.jsonl.gz",
-    pathlib.Path(floss.qs.db.__file__).parent / "data" / "gp" / "cwindb-dotnet.jsonl.gz",
-    pathlib.Path(floss.qs.db.__file__).parent / "data" / "gp" / "xaa-hashes.bin",
-    pathlib.Path(floss.qs.db.__file__).parent / "data" / "gp" / "yaa-hashes.bin",
+    _data_root() / "gp" / "gp.jsonl.gz",
+    _data_root() / "gp" / "cwindb-native.jsonl.gz",
+    _data_root() / "gp" / "cwindb-dotnet.jsonl.gz",
+    _data_root() / "gp" / "xaa-hashes.bin",
+    _data_root() / "gp" / "yaa-hashes.bin",
 )
