+++ pyproject.toml
@@ -103,6 +103,13 @@ dependencies = [
     # we still support.
 
     "networkx>=3",
+    # layout / tagging (formerly optional [qs])
+    "dnfile==0.13.0",
+    "colorama==0.4.6",
+    "machofile==2026.2.4",
+    "msgspec==0.14.2",
+    "python-lancelot==0.9.7",
+    "pyelftools==0.31",
 ]
 dynamic = ["version", "readme"]
 
@@ -118,14 +125,6 @@ packages = ["floss", "floss.sigs"]
 # we want all developer environments to be consistent.
 # These dependencies are not used in production environments
 # and should not conflict with other libraries/tooling.
-qs = [
-    "dnfile==0.13.0",
-    "colorama==0.4.6",
-    "machofile==2026.2.4",
-    "msgspec==0.14.2",
-    "python-lancelot==0.9.7",
-    "pyelftools==0.31",
-]
 dev = [
     "pre-commit==4.6.0",
     "pyyaml==6.0.1",
@@ -156,4 +155,3 @@ Signatures = "https://github.com/mandiant/flare-floss/tree/master/floss/sigs"
 
 [project.scripts]
 floss = "floss.main:main"
-qs = "floss.qs.main:main"
