+++ tests/test_gp_db.py
@@ -1,20 +1,20 @@
 import pathlib
 
-import floss.qs.db.gp
+import floss.tags.gp
 
 CD = pathlib.Path(__file__).parent
 
 
 def test_load_db():
-    path = pathlib.Path(floss.qs.db.gp.__file__).parent / "data" / "gp" / "gp.jsonl.gz"
-    db = floss.qs.db.gp.StringGlobalPrevalenceDatabase.from_file(path)
+    path = pathlib.Path(floss.tags.gp.__file__).resolve().parents[1] / "qs" / "db" / "data" / "gp" / "gp.jsonl.gz"
+    db = floss.tags.gp.StringGlobalPrevalenceDatabase.from_file(path)
 
     assert len(db) > 0  # 21 entries at time of writing
 
 
 def test_query_db():
-    path = pathlib.Path(floss.qs.db.gp.__file__).parent / "data" / "gp" / "gp.jsonl.gz"
-    db = floss.qs.db.gp.StringGlobalPrevalenceDatabase.from_file(path)
+    path = pathlib.Path(floss.tags.gp.__file__).resolve().parents[1] / "qs" / "db" / "data" / "gp" / "gp.jsonl.gz"
+    db = floss.tags.gp.StringGlobalPrevalenceDatabase.from_file(path)
     res = db.metadata_by_string["!This program cannot be run in DOS mode."]
 
     assert len(res) == 1
@@ -28,15 +28,15 @@ def test_query_db():
 
 
 def test_load_hash_db():
-    path = pathlib.Path(floss.qs.db.gp.__file__).parent / "data" / "gp" / "xaa-hashes.bin"
-    db = floss.qs.db.gp.StringHashDatabase.from_file(path)
+    path = pathlib.Path(floss.tags.gp.__file__).resolve().parents[1] / "qs" / "db" / "data" / "gp" / "xaa-hashes.bin"
+    db = floss.tags.gp.StringHashDatabase.from_file(path)
 
     assert len(db) > 0
 
 
 def test_query_hash_db():
-    path = pathlib.Path(floss.qs.db.gp.__file__).parent / "data" / "gp" / "xaa-hashes.bin"
-    db = floss.qs.db.gp.StringHashDatabase.from_file(path)
+    path = pathlib.Path(floss.tags.gp.__file__).resolve().parents[1] / "qs" / "db" / "data" / "gp" / "xaa-hashes.bin"
+    db = floss.tags.gp.StringHashDatabase.from_file(path)
 
     assert "!This program cannot be run in DOS mode." in db
     assert "Willi rules" not in db
