+++ tests/test_qs_elf.py
@@ -1,6 +1,7 @@
 from pathlib import Path
 
-from floss.qs.main import Slice, ELFLayout, SegmentLayout, compute_layout
+from floss.ranges import Slice
+from floss.layout import ELFLayout, SegmentLayout, compute_layout
 
 CD = Path(__file__).resolve().parent
 ELF_DIR = CD / "data" / "elf"
