+++ tests/test_winapi_db.py
@@ -1,17 +1,17 @@
 import pathlib
 
-import floss.qs.db.winapi
+import floss.tags.winapi
 
 
 def test_load_db():
-    path = pathlib.Path(floss.qs.db.winapi.__file__).parent / "data" / "winapi"
-    db = floss.qs.db.winapi.WindowsApiStringDatabase.from_dir(path)
+    path = pathlib.Path(floss.tags.winapi.__file__).resolve().parents[1] / "qs" / "db" / "data" / "winapi"
+    db = floss.tags.winapi.WindowsApiStringDatabase.from_dir(path)
     assert len(db) > 0
 
 
 def test_query_db():
-    path = pathlib.Path(floss.qs.db.winapi.__file__).parent / "data" / "winapi"
-    db = floss.qs.db.winapi.WindowsApiStringDatabase.from_dir(path)
+    path = pathlib.Path(floss.tags.winapi.__file__).resolve().parents[1] / "qs" / "db" / "data" / "winapi"
+    db = floss.tags.winapi.WindowsApiStringDatabase.from_dir(path)
 
     assert "kernel32.dll" in db.dll_names
     assert "kernel33.dll" not in db.dll_names
